% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_targetted.R
\name{generate_feature_t}
\alias{generate_feature_t}
\title{Generate feature data for targeted proteome.}
\usage{
generate_feature_t(ptm_site, flanking_size = 12, positive_info_file,
  known_protein_fasta_file, predict_protein_fasta_file, output_label_training,
  output_label_predict)
}
\arguments{
\item{ptm_site}{The amino acid this PTM involves, in upper-case single letter representation.}

\item{flanking_size}{The number of residues surround each side of the center residue, the total window size will be 2*flanking_size+1, default to 12.}

\item{positive_info_file}{A text file containing the positive PTM sites info in required format.}

\item{known_protein_fasta_file}{A fext file containing the proteins sequences of interest and known PTM sites in Fasta format.}

\item{predict_protein_fasta_file}{A fext file containing the proteins sequences with PTM sites to be predicted in Fasta format.}

\item{output_label_training}{The string to tag the output files associated with training proteins.}

\item{output_label_predict}{The string to tag the output files associated with prediction proteins.}
}
\description{
This function generates feature data for targeted by mapping the positive PTM info to protein sequences, constructing windows and extracting 3 sets of features.
}
\details{
This function outputs the features generated from input files.
}
\examples{
generate_feature_t(ptm_site = "S",
            flanking_size = 12,
            positive_info_file = "known_ps.tsv",
            known_protein_fasta_file = "known_fasta.tsv",
            predict_protein_fasta_file = "predict_fasta.tsv",
            output_label_training = "ps_training",
            output_label_predict = "ps_predict")
}
