% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_targetted.R
\name{generate_training_test_data_t}
\alias{generate_training_test_data_t}
\title{Generate training and prediction datasets for targeted proteome}
\usage{
generate_training_test_data_t(lower_bound = -1, upper_bound = 1,
  positive_training_feature_file, negative_training_feature_file,
  negative_predict_feature_file, output_label_training, output_label_predict)
}
\arguments{
\item{lower_bound}{The lower bound of the scaled data range, default to -1.}

\item{upper_bound}{The upper bound of the scaled data range, default to 1.}

\item{positive_training_feature_file}{A Rds file containing the positive training features(generated by g_feature_t()).}

\item{negative_training_feature_file}{A Rds file containing the negative training features(generated by g_feature_t()).}

\item{negative_predict_feature_file}{A Rds file containing the negative prediction features(generated by g_feature_t()).}

\item{output_label_training}{The string to tag the output files associated with training proteins.}

\item{output_label_predict}{The string to tag the output files associated with prediction proteins.}
}
\description{
This function generates training and prediction datasets for whole proteome to predict 1/n_fold of the proteome by the other n-1 fold data.
}
\details{
This function outputs formatted feature files ready for Liblinear training and prediction.
}
\examples{

generate_training_test_data_t(lower_bound = -1, upper_bound = 1,
positive_training_feature_file ="ps_training_noc_not_na_pos_feature.Rds",
negative_training_feature_file = "ps_training_noc_not_na_candi_feature.Rds",
negative_predict_feature_file = "ps_predict_noc_not_na_candi_feature.Rds",
output_label_training = "ps_training",
output_label_predict = "ps_predict")
}
