% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_whole_proteom.R
\name{generate_training_test_data_wp}
\alias{generate_training_test_data_wp}
\title{Generate training and prediction datasets for whole proteome}
\usage{
generate_training_test_data_wp(n_fold = 2, lower_bound = -1,
  upper_bound = 1, positive_feature_file, negative_feature_file, output_label)
}
\arguments{
\item{n_fold}{Number of folds used for training and prediction, default set to 2}

\item{lower_bound}{The lower bound of the scaled data range, default to -1.}

\item{upper_bound}{The upper bound of the scaled data range, default to 1.}

\item{positive_feature_file}{A Rds file containing the positive features(generated by g_feature_wp()).}

\item{negative_feature_file}{A Rds file containing the negative features(generated by g_feature_wp()).}

\item{output_label}{The string to tag the output files.}
}
\description{
This function generates training and prediction datasets for whole proteome to predict 1/n_fold of the proteome by the other n-1 fold data.
}
\details{
This function outputs formatted feature files ready for Liblinear training and prediction.
}
\examples{
generate_training_test_data_wp(n_fold = 2,
                        lower_bound = -1,
                        upper_bound = 1,
                        positive_feature_file = "ps_0103_noc_not_na_pos_feature.Rds",
                        negative_feature_file = "ps_0103_noc_not_na_candi_feature.Rds",
                        output_label = "ps_0103")
}
