% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_targetted.R
\name{predict_with_liblinear}
\alias{predict_with_liblinear}
\title{Process prediction with Liblinear}
\usage{
predict_with_liblinear(liblinear_dir, feature_file_path, training_file_name,
  test_file_name, output_label, cvlog_path_name)
}
\arguments{
\item{liblinear_dir}{Absolute path of Liblinear tool.}

\item{feature_file_path}{Absolute path of the feature files.}

\item{training_file_name}{A string indicating file name of the training feature data.}

\item{test_file_name}{A string indicating the file name of the test feature data.}

\item{output_label}{The string to tag the output files.}

\item{cvlog_path_name}{The path and name of the log files, which hold the details of Liblinear procedures.}
}
\description{
Process the feature files with Liblinear training and prediction. Train on known proteins and predict on unknown proteins
}
\details{
This function call Liblinear library to perform n_fold training/prediction, the prediction score will be part of the output files.
}
\examples{
predict_with_liblinear(liblinear_dir = "/data/ginny/liblinear-2.11/",
                      feature_file_path = "/data/ginny/test_package/",
                      training_file_name = "ps_predict_train_feature_svm.tsv",
                      test_file_name = "ps_predict_test_feature_svm.tsv",
                      output_label = "ps_predict",
                      cvlog_path_name = "/data/ginny/test_package/cvlog.txt")
}
