% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_whole_proteom.R
\name{present_prediction_wp}
\alias{present_prediction_wp}
\title{Present Liblinear prediction results}
\usage{
present_prediction_wp(positive_index_file_names, candi_index_file_names,
  prediction_score_file_names, test_label_file_names, candidate_df_Rds,
  specificity_level = 0.99, output_label)
}
\arguments{
\item{positive_index_file_names}{An Rds file containing the names of indices for positive windows.}

\item{candi_index_file_names}{An Rds file containing the names of indices for candidate(negative) windows.}

\item{prediction_score_file_names}{An Rds file containing the file names of Liblinear predicted scores.}

\item{test_label_file_names}{An Rds file containing the file names of the label of the test data.}

\item{candidate_df_Rds}{An Rds file containing the data frame of all candiate sites information(generated by g_feature_wp)/}

\item{specificity_level}{A numerical number indicating the specificity user requires the classifier to achieve, default set to 0.99}

\item{output_label}{The string to tag the output files.}
}
\description{
Combine Liblinear prediction results and window, position informations of sites. Calculate score threshold at user specified sensitivity level.
}
\examples{
present_prediction_whole_proteome(positive_index_file_names = "ps_0103_pos_index_names.Rds",
                                 candi_index_file_names = "ps_0103_candi_index_names.Rds",
                                 prediction_score_file_names = "ps_0103_prediction_file_names.Rds",
                                 test_label_file_names = "ps_0103_test_label_names.Rds",
                                 candidate_df_Rds = "ps_0103_candidate.Rds",
                                 specificity_level = 0.99,
                                 output_label = "ps_0103")
}
