% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_whole_proteom.R
\name{process_n_fold_cross_validation}
\alias{process_n_fold_cross_validation}
\title{Process n_fold cv}
\usage{
process_n_fold_cross_validation(liblinear_dir, n_fold, feature_file_path,
  training_file_names, test_file_names, output_label, cvlog_path_name)
}
\arguments{
\item{liblinear_dir}{Absolute peth of Liblinear tool.}

\item{n_fold}{Number of folds used for training and prediction, default set to 2}

\item{feature_file_path}{Absolute path of the feature files.}

\item{training_file_names}{An Rds file containing the file names of the training feature data.}

\item{test_file_names}{An Rds file containing the file names of the test feature data.}

\item{output_label}{The string to tag the output files.}

\item{cvlog_path_name}{The path and name of the log files, which hold the details of Liblinear procedures.}
}
\description{
Process the feature files with Liblinear training and prediction. Conduct n_fold cross validation with the files supplied.
}
\details{
This function call Liblinear library to perform n_fold training/prediction, the prediction score will be part of the output files.
}
\examples{
process_n_fold_cross_validation(liblinear_dir = "/data/ginny/liblinear-2.11/",
                                n_fold = 2,
                                feature_file_path = "/data/ginny/test_package/",
                                training_file_names = "ps_0103_training_feature_names.Rds",
                                test_file_names = "ps_0103_test_feature_names.Rds",
                                output_label = "ps_0103",
                                cvlog_path_name = "/data/ginny/test_package/cvlog.txt")

}
