% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_domain.R
\name{PTM_domain_mapping_enrichment}
\alias{PTM_domain_mapping_enrichment}
\title{Building up domain analysis.}
\usage{
PTM_domain_mapping_enrichment(window_score_label_Rds, output_label)
}
\arguments{
\item{window_score_label_Rds}{An Rds file containing the window, positive and score of interested sites(genereated from present_predict_wp()).}

\item{output_label}{The string to tag the output files.}
}
\description{
Map sites to protein domains and analyse the enrichment of PTM sites for each domain.
}
\examples{
PTM_domain_mapping_enrichment(window_score_label_Rds = "ps_0103_window_score_df.Rds",
                              output_label = "ps_0103")
}
