% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_enrichment.R
\name{calculate_tbt_positive_ptms}
\alias{calculate_tbt_positive_ptms}
\title{Analyse positive cross talk events in domains.}
\usage{
calculate_tbt_positive_ptms(distance = 5, anchor_mod, cross_mod,
  anchor_mapped_df_Rds, cross_mapped_df_Rds, output_label)
}
\arguments{
\item{distance}{A numerical value indicating distance between two PTM sites, defaul set to 5.}

\item{anchor_mod}{A string indicating the anchor modification.}

\item{cross_mod}{A string indicating the crosstalk modification.}

\item{anchor_mapped_df_Rds}{An Rds file containing the achor window score file with domain mapped(generated by PTM_domain_mapping_enrichment())}

\item{cross_mapped_df_Rds}{An Rds file containing the cross window score file with domain mapped(generated by PTM_domain_mapping_enrichment())}

\item{output_label}{The string to tag the output files.}
}
\description{
Test for enrichment of two user provided PTM types inside of domains.
}
\examples{
calculate_tbt_positive_ptms(distance = 5,
                       anchor_mode = "ps",
                       cross_mod = "pt",
                       anchor_mapped_df_Rds = "ps_0103_mapped_df.Rds",
                       cross_mapped_df_Rds = "pt_0103_mapped_df.Rds",
                       output_label = "pt_ps_positive")
}
