% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_targetted.R
\name{predict_on_targeted_proteome}
\alias{predict_on_targeted_proteome}
\title{Predict on targeted proteome.}
\usage{
predict_on_targeted_proteome(ptm_site, flanking_size = 12, positive_info_file,
  known_protein_fasta_file, predict_protein_fasta_file, output_label_training,
  output_label_predict, lower_bound = -1, upper_bound = 1, liblinear_dir,
  feature_file_path, cvlog_path_name, specificity_level,
  flag_for_score_threshold_chosen = "reference", score_threshold)
}
\arguments{
\item{ptm_site}{The amino acid this PTM involves, in upper-case single letter representation.}

\item{flanking_size}{The number of residues surround each side of the center residue, the total window size will be 2*flanking_size+1, default to 12.}

\item{positive_info_file}{A text file containing the positive PTM sites info in required format.}

\item{known_protein_fasta_file}{A fext file containing the proteins sequences of interest and known PTM sites in Fasta format.}

\item{predict_protein_fasta_file}{A fext file containing the proteins sequences with PTM sites to be predicted in Fasta format.}

\item{output_label_training}{The string to tag the output files associated with training proteins.}

\item{output_label_predict}{The string to tag the output files associated with prediction proteins.}

\item{lower_bound}{the lower bound of the scaled data range, default to -1.}

\item{upper_bound}{the upper bound of the scaled data range, default to 1.}

\item{liblinear_dir}{Absolute peth of Liblinear tool.}

\item{feature_file_path}{Absolute path of the feature files.}

\item{cvlog_path_name}{The path and name of the log files, which hold the details of Liblinear procedures.}

\item{specificity_level}{A numerical number indicating the specificity user requires the classifier to achieve, default set to 0.99. Used only not in "reference" mode.}

\item{flag_for_score_threshold_chosen}{A string indicating whether use reference score threshold or get from the user supplied training data, defalt set to "reference".}

\item{score_threshold}{A numerical value between 0 to 1 indicating the reference score threshold (supply when in "reference")}

\item{n_fold}{Number of folds used for training and prediction, default set to 2}
}
\description{
Predict on targeted proteome.
}
\details{
This function outputs the features generated from input files.
}
\examples{
predict_on_targeted_proteome = function (ptm_site = "S", 
                                         flanking_size=12, 
                                         positive_info_file = "known_ps.tsv", 
                                         known_protein_fasta_file = "known_fasta.tsv",
                                         predict_protein_fasta_file = "predict_fasta.tsv",
                                         output_label_training = "ps_training",
                                         output_label_predict = "ps_predict",
                                         lower_bound = -1,
                                         upper_bound = 1,
                                         liblinear_dir = "/data/ginny/liblinear-2.11/",
                                         feature_file_path = "/data/ginny/test_package/",
                                         cvlog_path_name = "/data/ginny/test_package/cvlog.txt",
                                         specificity_level = 0.99,
                                         flag_for_score_threshold_chosen = "cv",
                                         score_threshold = NULL)
}
