% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_whole_proteom.R
\name{predict_on_whole_proteome}
\alias{predict_on_whole_proteome}
\title{A function to predict and annotate whole proteom PTM sites}
\usage{
predict_on_whole_proteome(ptm_site, flanking_size = 12, positive_info_file,
  protein_fasta_file, n_fold = 2, lower_bound = -1, upper_bound = 1,
  liblinear_dir, feature_file_path, cvlog_path_name, specificity_level = 0.99,
  output_label)
}
\arguments{
\item{ptm_site}{The amino acid this PTM involves, in upper-case single letter representation.}

\item{flanking_size}{The number of residues surround each side of the center residue, the total window size will be 2*flanking_size+1, default to 12.}

\item{positive_info_file}{A text file containing the positive PTM sites info in required format.}

\item{protein_fasta_file}{A fext file containing the proteins sequences of interest in Fasta format.}

\item{n_fold}{Number of folds used for training and prediction, default set to 2}

\item{lower_bound}{the lower bound of the scaled data range, default to -1.}

\item{upper_bound}{the upper bound of the scaled data range, default to 1.}

\item{liblinear_dir}{Absolute peth of Liblinear tool.}

\item{feature_file_path}{Absolute path of the feature files.}

\item{cvlog_path_name}{The path and name of the log files, which hold the details of Liblinear procedures.}

\item{specificity_level}{A numerical number indicating the specificity user requires the classifier to achieve, default set to 0.99.}

\item{output_label}{The string to tag the output files.}
}
\description{
A function to predict and annotate whole proteom PTM sites
}
\details{
This function outputs the features generated from input files.
}
\examples{
predict_on_whole_proteome(ptm_site = "S",
                         flanking_size = 12,
                         positive_info_file = "ps_PSP.tsv",
                         protein_fasta_file = "S_sp_fasta.tsv",
                         n_fold = 2,
                         lower_bound = -1,
                         upper_bound = 1,
                         liblinear_dir = "/data/ginny/liblinear-2.11/",
                         feature_file_path = "/data/ginny/test_package/",
                         cvlog_path_name = "/data/ginny/test_package/cvlog.txt",
                         specificity_level = 0.99,
                         output_label = "ps_0103")
}
